Embed roundcube into nextcloud.
This commit is not a "nice" commit w.r.t. to the project used for
forking. However, the changes are quite large, so this is rather a harsh
break. Some features for the original App by Martin Reinhardt will be
reintroduced (concerning the login scheme).

In particular the "user-id must be an email-address" features is just a no-go for me.